refactor: move discovery query/filter helpers to cli/commands/discovery.py
